omit SPDX expressions from exported list
The SPDX JSON file at https://spdx.org/licenses/licenses.json currently
includes both SPDX identifiers like "GPL-3.0" for "the GNU Public
License version 3.0" and some expressions with the plus operator, like
"GPL-3.0+" for "the GNU Public License version 3.0 or later". The latter
is technically an SPDX expression with the identifier "GPL-3.0" and the
"+" operator, and not an SPDX identifier.

This patch filters the elements of the licenses array in SPDX.org's JSON
file to omit listings that end with "+", to omit those elements that are
actually SPDX expressions with the "+" SPDX expression operator.